Se agrega el schema y nodo para los productos, además, de una bd en formato json con productos que se precargaran por defecto despues de crea la migración

+++ back-end/apps/products/models.py
@@ -7,7 +7,7 @@ class Product(models.Model):
     # Nombre del producto
     name = models.CharField(max_length = 200)
     # Valor del producto
-    value = models.DecimalField(max_digits = 10, decimal_places = 2)
+    value = models.IntegerField()
     # Breve descripcion del producto
     description = models.TextField(blank = True)
     # Nombre de la imagen almacenada en S3
